+++ packages/examples/mongodb-express/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @better-conversation/example-mongodb-express
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @better-conversation/adapter-mongodb@1.0.2
+
 ## 0.0.2
 
 ### Patch Changes